+++ src/lib/math.h
@@ -3,4 +3,4 @@
 #include <types.h>
 
 i32 abs(i32 i);
-u64 jdn(u8 days, u8 months, u8 years);
+u64 jdn(u8 days, u8 months, u16 years);
